test(sensitive-data): fix tests according to npm version
